Move methods to more specific classes
