+++ app/views/auth0/pre_login.html.erb
@@ -1,4 +1,6 @@
-<h1>Login using auth0</h1>
-<img src="https://cdn.auth0.com/styleguide/1.0.0/img/badge.svg">
+<h1>IDseq auth0</h1>
+<img src="https://cdn.auth0.com/styleguide/1.0.0/img/badge.svg" width="100" height="100"/>
 
-<%= button_to 'Auth0', '/auth/auth0', method: :post %>
+<p style="color: red"><%= @error_message %></p>
+
+<%= button_to 'Login using auth0', '/auth/auth0', method: :post %>
